Fixed urls not working on Android 11
Removed can launch since `http` and `https` uri's are always allowed to be launched.

url_launcher uses a package query to check if there are apps that can launch the `uri`. But since `http` and `https` uri's are (almost) always able to be opened, this has become useless. There also isn't  a query that I could find for `https` and/ or `http` support in the manifest.

Removed `<uses-permission android:name"android.permission.QUERY_ALL_PACKAGES"/>` from the Android manifest since the app didn't match the criteria on https://developer.android.com/training/basics/intents/package-visibility#all-apps with the risk of the Google Play store rejecting the update if left in.