Fixed `groups` table not being renamed in the database migration
In older version of drift the `groups` table was named `"groups"` and in newer versions it has been renamed to `groups` so it needs to be migrated.

Updated the migration tests for these changes. The migration tests now also take in to account that the `groups` table has a different name in for the scheme version 3.